+++ neo/idlib/CmdArgs.cpp
@@ -155,7 +155,7 @@ void idCmdArgs::TokenizeString( const char *text, bool keepAsStrings ) {
 
 		len = token.Length();
 
-		if ( totalLen + len + 1 > sizeof( tokenized ) ) {
+		if ( (unsigned int)(totalLen + len + 1) > sizeof( tokenized ) ) {
 			return;			// this is usually something malicious
 		}
